requirements.txt
duckdb==1.2.0
ipykernel==6.29.5
matplotlib==3.10.1
matplotlib-inline==0.1.7
pandas==2.2.3
python-dotenv==1.0.1
requests==2.32.3
seaborn==0.13.2